minor bug fixes

+++ src/controllers/MeetingController.js
@@ -3,7 +3,6 @@ const {User, Meeting, MeetingMember, CandidateVote, Vote} = require('../models')
 module.exports = {
     async createGuestMeeting(req, res) {
         try {
-            console.log(req.body);
             const user = await User.create({
                 name: req.body.name
             });
@@ -21,32 +20,29 @@ module.exports = {
             });
             res.send();
         } catch (error) {
-            console.log(error);
             res.status(500).send({
                 error: 'Something went wrong with creating a meeting, please try again later'
             })
         }
     },
     async addGuestUser(req, res) {
-      try {
-          const user = await User.create({
-             name: req.body.name
-          });
-          await MeetingMember.create({
-             userID: user.id,
-             meetingID: req.body.meetingID
-          });
-          res.status(200).send();
-      } catch (error) {
-          console.log()
-          res.status(500).send({
-              error: 'Something went wrong with adding a user to the meeting, please try again later'
-          })
-      }
+        try {
+            const user = await User.create({
+                name: req.body.name
+            });
+            await MeetingMember.create({
+                userID: user.id,
+                meetingID: req.body.meetingID
+            });
+            res.status(200).send();
+        } catch (error) {
+            res.status(500).send({
+                error: 'Something went wrong with adding a user to the meeting, please try again later'
+            })
+        }
     },
     async editMeetingType(req, res) {
         try {
-            console.log(req.body.pollType)
             await Meeting.update({
                 pollType: req.body.pollType,
             }, {
@@ -54,9 +50,8 @@ module.exports = {
                     meetingID: req.body.meetingID
                 }
             });
-            res.send();
+            res.status(200).send();
         } catch (error) {
-            console.log()
             res.status(500).send({
                 error: 'Something went wrong with editing the meeting type, please try again later'
             })
@@ -72,9 +67,8 @@ module.exports = {
                     meetingID: req.body.meetingID
                 }
             });
-            res.send();
+            res.status(200).send();
         } catch (error) {
-            console.log()
             res.status(500).send({
                 error: 'Something went wrong with editing meeting details, please try again later'
             });
@@ -110,21 +104,20 @@ module.exports = {
                         }
                     })
                 }
-                // todo check if user is a guest user, if it is then delete
-                await User.destroy({
-                    where: {
-                        id: users[i].id
-                    }
-                });
+                if (users[i].guest === true)
+                    await User.destroy({
+                        where: {
+                            id: users[i].id
+                        }
+                    });
             }
             await Meeting.destroy({
                 where: {
                     meetingID: req.body.meetingID
                 }
             });
-            res.send("success");
+            res.status.send();
         } catch (error) {
-            console.log(error);
             res.status(500).send({
                 error: 'Something went wrong with deleting the meeting, please try again later'
             });
